Begin implementation of a user account system
Create a class to manage users, while removing nick. There is no implementation of the class with anything else, like through the chat, but it should be trivial for someone who actually knows how those work to implement them. There are also no bindings to the database yet, so all of the users will sit in memory, which will be fixed later. If you have to fix something, you have to break it first :-)