Close #63: set name client side.
This allows players to define a name in a configuration file. The
name is sent to the server as the first thing after connecting. If
no name is given or if the name is the default "[unnamed]" the
server will default to the old "player%d" name.

Does not prevent name collision.