vulkanize: Convert uniforms to ubos
UBOs are required for SPIR-V. This converts the list of uniforms in each
shader to a uniform block that is backed by a buffer object that is
updated accordingly. This requires a version bump in the shaders.
Samplers are an exception and they continue to be separate.

Since UBO representation differs by API, the UBO is wrapped in a Uniform
interface object that will eventually include the samplers as well.
The uniform interface object is bound alongside the attrib object.