vulkanize: Split text pipelines and shaders
Because blending state in Vulkan is immutable and text rendering and
sign rendering use different blending states, they will need to be
separated. Its ubo was needlessly distinct due to a single
boolean that split the behavior of the fragment shader in almost every
respect. It makes more sense for them to be split.