vulkanize: Separate drawing functions
To facilitate replacing GL calls with Vulkan equivalents, functions that
call GL code or are closely related to those that do are moved into
glrenderer files. A few functions had their prototypes changed to remove
depdendencies on types that are exclusive to main.c. Attrib is closely
related to the renderer, so glrenderer takes ownership of it.

GL calls in functions that are not to be moved have been left for now,
but will be moved in a later change.