vulkanize: Abstract textures and samplers
Collects texture information into an Image object and sampler
information into a Uniform object. Both are bound through a
bind_pipeline call that takes the pre-existing Attrib object as well.

Shortly, the samplers in the uniform interface object will be joined by
UBOs to encompass the other uniforms.

Additionally, texture samplers are now bound explicitly and tightly
within the shaders themselves. To use this feature, the version had to
be bumped to 420. This is among the alterations needed to compile them
to Spir-V