vulkanize: Rename pipeline and abstract contents
Since all that's left in what was the Attrib object is a program, it
really doesn't make sense to call it that anymore. It will be used as a
program/pipeline object going forward. So it's named Pipeline.
Additionally, the struct has been rendered undefined outside the
renderer code so it's contents are API depdenent.