vulkanize: Introduce updatable dynamic buffer
Rather than delete each buffer after its done and recreate it, we can
create dynamic buffers for things with attributes that change rapidly.
With them, some update functions that preserve the backing store of the
VBO and just update the contents.

This describes what is done currently, with vulkan, they will consist of
multiple objects, one for each frame in flight, but the interface will
be the same.