vulkanize: Enable split screen viewing
The first part of this is to implement setting the viewport, which was
deferred before. This means that the viewport must be dynamic for every
pipeline. Okay. not every one, but it's easier to make them the same.

Much of this involves adding uniforms for the portions of the scene that
are rendered in the split screen. Since it is derived from a different
player, the values are entirely independent and need their own uniforms,
but the pipelines can be reused.

The text buffer is considerably lesser on the split screen, but it's
simpler and not terribly inefficient to use the same mechanism.

Like so many others, player vertex buffers were getting created and
deleted on a per-frame basis. This converts them to dynamic buffers and
updates them instead of recreating them. A side effect of this is that
the CPU finally caught up with the v-synch and revealed that I left out
a fence wait at start_frame.