Add file to translate image into integer array
Takes an image file as input an returns the rgb array as a integer array of values 0 to 100. 0 being black and 100 being white. This was done because the images need to be turned into an integer array to be transformed later into an amount of blocks based on a grayscale.